% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_codes.R
\docType{data}
\name{data_codes}
\alias{data_codes}
\title{Diagnoses Code Descriptions}
\format{
A data frame with columns:
\describe{
  \item{DG}{Diagnosis code}
  \item{CODECLASS}{ICD Diagnose Class: ICD10, ICD9 or ICD8}
  \item{CATEGORY}{Summary Level of Diagnose Category}
  \item{DESC}{Description of the Diagnosis}
}
}
\source{
Internal project curation / THL registry / etc.
}
\usage{
data_codes
}
\description{
A dataset containing mappings of diagnosis codes to their descriptions.
}
\keyword{datasets}
