% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_clean.R
\name{.regex_clean}
\alias{.regex_clean}
\title{Clean Diagnosis Regex Strings}
\usage{
.regex_clean(dglist)
}
\arguments{
\item{dglist}{character. A string or vector of diagnosis code patterns.}
}
\value{
A character vector with all spaces removed and converted to uppercase.
}
\description{
Prepares diagnosis code regular expressions by removing spaces and converting to uppercase.
}
\examples{
regex_clean("e11 ")
regex_clean(c(" i10", " E11 "))

}
