% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_icd10_3letters.R
\name{recode_icd10_3letters}
\alias{recode_icd10_3letters}
\title{Recode diagnosis code to ICD-10 three-letter format}
\usage{
recode_icd10_3letters(DG, DGREG = DGREG)
}
\arguments{
\item{DG}{A character vector of diagnosis codes.}

\item{DGREG}{A character scalar or vector indicating the diagnosis coding system. Expected value is "ICD10".}
}
\value{
A character vector of the first three letters of ICD-10 codes, or `NA` if `DGREG` is not "ICD10".
}
\description{
Extracts the first three characters from an ICD-10 diagnosis code if the coding system is ICD-10.
}
\examples{
diagnoses <- diagnoses_raw |>
 dplyr::mutate(
   ICD10_3LETTERS = recode_icd10_3letters(DG, DGREG = DGREG),
 )

}
