% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_classification.R
\name{classify_population}
\alias{classify_population}
\title{Create Exposure and Optional Response Groups from Diagnoses}
\usage{
classify_population(
  exposure_icd10 = "",
  exposure_icd9 = "",
  exposure_icd8 = "",
  exposure_src = c("avohilmo", "erko", "hilmo", "local", "ksyy", "soshilmo", "syopa"),
  response_icd10 = NULL,
  response_icd9 = NULL,
  response_icd8 = NULL,
  response_src = c("avohilmo", "erko", "hilmo", "local", "ksyy", "soshilmo", "syopa"),
  data_population = population,
  data_diagnoses = diagnoses,
  runtime_shiny = FALSE
)
}
\arguments{
\item{exposure_icd10}{A character vector of ICD-10 codes (regex-supported) to identify exposure group.}

\item{exposure_icd9}{A character vector of ICD-9 codes (regex-supported) to identify exposure group.}

\item{exposure_icd8}{A character vector of ICD-8 codes (regex-supported) to identify exposure group.}

\item{exposure_src}{A character vector of registry sources used to search for exposure diagnoses.
Defaults to all available sources.}

\item{response_icd10}{Optional character vector of ICD-10 codes for response group (default: NULL).}

\item{response_icd9}{Optional character vector of ICD-9 codes for response group (default: NULL).}

\item{response_icd8}{Optional character vector of ICD-8 codes for response group (default: NULL).}

\item{response_src}{A character vector of registry sources used to search for response diagnoses.
Defaults to all available sources.}

\item{data_population}{A data frame of the target population (default: `population`).}

\item{data_diagnoses}{A data frame of diagnoses (default: `diagnoses`).}

\item{runtime_shiny}{Logical; if TRUE and run inside a Shiny app, shows progress bar (default: FALSE).}
}
\value{
A data frame with exposure (and optionally response) group classifications and metadata,
        including ID, birth/death/migration dates, diagnosis info, and binary indicators for exposure/response.
}
\description{
This function identifies individuals with exposure diagnoses (and optionally response diagnoses)
based on ICD-10/9/8 codes and selected registry sources. It returns a dataset of grouped individuals
with exposure (and optionally response) classifications, diagnosis dates, and relevant metadata.
Function searches first diagnose date, which is given for exposure/response diagnose date.
}
\examples{
classify_population(exposure_icd10 = "I21", exposure_src = "hilmo")
classify_population(exposure_icd10 = "F32", response_icd10 = "I21")

}
