% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{plot_survival_km}
\alias{plot_survival_km}
\title{Plot Kaplan-Meier Survival Curve (Exposure to Response)}
\usage{
plot_survival_km(
  data,
  color = "#D9534F",
  scale = "years",
  plot = c("base", "survminer")
)
}
\arguments{
\item{data}{A data frame in long format with the following required columns:
\describe{
  \item{ID}{Subject ID}
  \item{name}{Type of event: must include `"diagnose"` for response; other values are treated as censored}
  \item{value}{Time (in days) from exposure to event}
}}

\item{scale}{years or days (timeline)}
}
\value{
A Kaplan-Meier survival curve plotted using the base R `plot()` function.
}
\description{
Plots an overall Kaplan-Meier survival curve from a long-format survival dataset
             containing response, censoring, and death events.
}
\details{
Internally converts event types to binary (1 = diagnose, 0 = other), constructs a `Surv` object,
and fits an overall survival model using `survfit()`. Designed to be used interactively in Shiny with
optional progress feedback via `.safe_inc_progress()` and `withProgress()`.
}
\examples{
\dontrun{
km_plot <- plot_survival_km(data = long_surv_df)
}

}
