% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_handling.R
\name{recode_icd10_class}
\alias{recode_icd10_class}
\title{Recode ICD-10 3-letter codes into broad classification groups}
\usage{
recode_icd10_class(ICD10_3LETTERS, DGREG = DGREG)
}
\arguments{
\item{ICD10_3LETTERS}{A character vector of 3-letter ICD-10 codes.}

\item{DGREG}{A character scalar or vector indicating the diagnosis coding system. Expected value is "ICD10".}
}
\value{
A character vector indicating the ICD-10 classification range, or `NA` if `DGREG` is not "ICD10".
}
\description{
Maps 3-letter ICD-10 codes into standard ICD-10 chapter ranges (e.g., "A00–B99", "C00–D48").
}
\examples{
diagnoses <- diagnoses_raw |>
 dplyr::mutate(
   ICD10_3LETTERS = recode_icd10_3letters(DG, DGREG = DGREG),
   ICD10_CLASS = recode_icd10_class(ICD10_3LETTERS, DGREG = DGREG)
 )

}
